# btf: lazy decoding