--- a/btf/strings_test.go
+++ b/btf/strings_test.go
@@ -102,15 +102,3 @@ func BenchmarkStringTableZeroLookup(b *testing.B) {
 		}
 	}
 }
-
-func newStringTable(strings ...string) *stringTable {
-	offsets := make([]uint32, len(strings))
-
-	var offset uint32
-	for i, str := range strings {
-		offsets[i] = offset
-		offset += uint32(len(str)) + 1 // account for NUL
-	}
-
-	return &stringTable{nil, offsets, 0, strings}
-}