Prompted by #1755 another stab at lazy decoding of BTF types. IterateVmlinux is roughly the pwru workload (IIRC) and InspectorGadget does what it says on the tin.

    core: 1
    goos: linux
    goarch: amd64
    pkg: github.com/cilium/ebpf/btf
    cpu: 13th Gen Intel(R) Core(TM) i7-1365U
                    │  base.txt   │          eager-strings.txt           │
                    │   sec/op    │   sec/op     vs base                 │
    ParseVmlinux      80.59m ± 4%   26.95m ± 2%  -66.56% (p=0.002 n=6)
    IterateVmlinux    193.7m ± 5%   151.8m ± 2%  -21.65% (p=0.002 n=6)
    InspectorGadget   87.22m ± 1%   34.56m ± 2%  -60.37% (p=0.002 n=6)
    BuildVmlinux      145.7m ± 4%
    geomean           118.7m        52.09m       -53.00%               ¹
    ¹ benchmark set differs from baseline; geomeans may not be comparable
    
                    │   base.txt    │           eager-strings.txt           │
                    │     B/op      │     B/op      vs base                 │
    ParseVmlinux      24.427Mi ± 0%   9.970Mi ± 0%  -59.18% (p=0.002 n=6)
    IterateVmlinux     52.51Mi ± 0%   34.71Mi ± 0%  -33.90% (p=0.002 n=6)
    InspectorGadget    26.65Mi ± 0%   12.00Mi ± 0%  -54.97% (p=0.002 n=6)
    BuildVmlinux       42.22Mi ± 0%
    geomean            34.66Mi        16.07Mi       -50.47%               ¹
    ¹ benchmark set differs from baseline; geomeans may not be comparable
    
                    │  base.txt   │          eager-strings.txt           │
                    │  allocs/op  │  allocs/op   vs base                 │
    ParseVmlinux      271.9k ± 0%   146.1k ± 0%  -46.27% (p=0.002 n=6)
    IterateVmlinux    398.9k ± 0%   272.9k ± 0%  -31.60% (p=0.002 n=6)
    InspectorGadget   280.8k ± 0%   155.0k ± 0%  -44.80% (p=0.002 n=6)
    BuildVmlinux      1.508k ± 0%
    geomean           82.32k        183.5k       -41.24%               ¹
    ¹ benchmark set differs from baseline; geomeans may not be comparable