@brb 2025-04-28
@lmb Are you interested in perf benchmarks, or correctness? If the latter, then all good - https://github.com/cilium/pwru/pull/548. Unfortunately, we do not have any benchmarking suites.