@lmb 2025-04-28
I added a crude benchmark which just iterates all of vmlinux to simulate pwru. Even that has gotten faster!